Properly sign extend the immediate value for some I-Type instructions
Previously, I was simply casting the bits to a signed int (reinterpreting them in two's complement)
However, that is not correct.
All bits in the u32 that are not part of the value must be set to the signed bit. (Read code for more detail)
This was causing multiple bitop tests to fail
Fixed:
* ORI
* ANDI
* XORI
(Note that most bitop instructions are not implemented yet)j